wire: wire.FieldsOf should not provide pointer to field type for non-pointer structs
