+++ internal/wire/testdata/FieldsOfStruct/foo/wire.go
@@ -26,10 +26,3 @@ func injectedMessage() string {
 		wire.FieldsOf(new(S), "Foo"))
 	return ""
 }
-
-func injectedMessagePtr() *string {
-	wire.Build(
-		provideS,
-		wire.FieldsOf(new(S), "Foo"))
-	return nil
-}
