+++ internal/wire/testdata/FieldsOfStruct/want/program_out.txt
@@ -1,2 +1,1 @@
 Hello, World!
-pointer to Hello, World!
