+++ internal/wire/testdata/FieldsOfStructDoNotProvidePtrToField/pkg
@@ -1,0 +1,1 @@
+example.com/foo
