+++ internal/wire/testdata/FieldsOfStructPointer/foo/foo.go
@@ -26,4 +26,5 @@ func provideS() *S {
 
 func main() {
 	fmt.Println(injectedMessage())
+	fmt.Println("pointer to " + *injectedMessagePtr())
 }
