+++ internal/wire/testdata/FieldsOfStructPointer/want/program_out.txt
@@ -1,1 +1,2 @@
 Hello, World!
+pointer to Hello, World!
