+++ internal/wire/testdata/FieldsOfStructPointer/want/wire_gen.go
@@ -12,3 +12,9 @@ func injectedMessage() string {
 	string2 := s.Foo
 	return string2
 }
+
+func injectedMessagePtr() *string {
+	s := provideS()
+	string2 := &s.Foo
+	return string2
+}
